+++ apps/svelte-manga/src/lib/components/ReadingProgress.svelte
@@ -1,17 +1,18 @@
 <!-- ReadingProgress.svelte -->
 
 <script lang="ts">
-	import { pb, getPocketbase } from '$lib/utils/api';
+	import { getPocketbase } from '$lib/utils/api';
 	import { onMount } from 'svelte';
 	import Icon from '@iconify/svelte';
+	import { page } from '$app/stores';
 
 	let readingProgress: any = [];
 
 	onMount(async () => {
-		if (pb.authStore.isValid) {
+		if ($page.data.user) {
 			const data = {
 				sort: '-updated',
-				filter: `user="${pb.authStore.model?.id}"`,
+				filter: `user="${$page.data.user?.id}"`,
 				expand: 'manga, currentChapter, user'
 			};
 			const res = await getPocketbase('reading_progress', data);
@@ -35,7 +36,7 @@
 		Your Reading Progress
 	</h2>
 
-	{#if pb.authStore.isValid}
+	{#if $page.data.user}
 		{#if readingProgress.length != 0}
 			<!-- Individual Chapters -->
 			<ul class="grid grid-cols-1 gap-4">