+++ apps/svelte-manga/src/lib/utils/api.ts
@@ -1,9 +1,13 @@
 import PocketBase from 'pocketbase';
 import { authData } from '$lib/utils/stores';
 import { goto } from '$app/navigation';
-import { site } from '$lib/config/site';
+import { site } from '@valiantlynx/general-config';
 
-export const pb = new PocketBase(site.pocketbase);
+export const pb = new PocketBase(site.site.pocketbase);
+
+export const getImageURL = (collectionId: string, recordId: string, fileName: string) => {
+	return `${site.site.pocketbase}/api/files/${collectionId}/${recordId}/${fileName}`;
+};
 
 export const authPocketbase = async (user: string, password: string) => {
 	const res = await pb.collection('users').authWithPassword(user, password);