+++ packages/svelte-ui/src/lib/index.js
@@ -1,0 +1,4 @@
+// Reexport your entry components here
+export { default as Input } from './components/Input.svelte';
+export { default as Load } from './components/Load.svelte';
+export { default as Modal } from './components/Modal.svelte';