+++ packages/svelte-ui/src/routes/+page.svelte
@@ -1,0 +1,3 @@
+<h1>Svelte ui components</h1>
+
+